TrainingList moved to new package "list"
