+++ app/src/main/AndroidManifest.xml
@@ -7,7 +7,8 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.TrainingTimer">
+        android:theme="@style/Theme.TrainingTimer"
+        android:name=".data.App">
         <activity
             android:name=".ui.MainActivity"
             android:windowSoftInputMode="adjustResize"
