+++ app/src/main/java/com/trainingtimer/ui/list/TrainingListFragment.kt
@@ -1,4 +1,4 @@
-package com.trainingtimer.ui
+package com.trainingtimer.ui.list
 
 import android.os.Bundle
 import android.util.Log
