+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/trainingListFragment"
-        android:name="com.trainingtimer.ui.TrainingListFragment"
+        android:name="com.trainingtimer.ui.list.TrainingListFragment"
         android:label="TrainingListFragment"
         tools:layout="@layout/fragment_training_list" >
